Share trust guidance without importing the command engine
